Bug 1967104 - Factor out the light-dark() function. r=dshin
Remove the light-dark() pref while at it.

Differential Revision: https://phabricator.services.mozilla.com/D249787
